% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaplot-package.R
\docType{package}
\name{areaplot-package}
\alias{areaplot-package}
\title{Plot Stacked Areas and Confidence Bands as Filled Polygons}
\description{
Plot stacked areas and confidence bands as filled polygons, or add polygons
to existing plots. A variety of input formats are supported, including
vectors, matrices, data frames, formulas, etc.
}
\details{
\emph{Plot:}
\tabular{ll}{
  \code{\link{areaplot}} \tab stacked area\cr
  \code{\link{confplot}} \tab confidence band
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arni-magnusson/areaplot}
}

}
\author{
Arni Magnusson.
}
